+++ facebook_auth/android/src/main/java/app/meedu/flutter_facebook_auth/FacebookLoginResultDelegate.java
@@ -2,17 +2,32 @@ package app.meedu.flutter_facebook_auth;
 
 
 import android.content.Intent;
+import android.util.Log;
+
+import com.facebook.AccessToken;
 import com.facebook.CallbackManager;
+import com.facebook.FacebookAuthorizationException;
 import com.facebook.FacebookCallback;
 import com.facebook.FacebookException;
+import com.facebook.login.LoginManager;
 import com.facebook.login.LoginResult;
+
 import java.util.HashMap;
+
 import io.flutter.plugin.common.MethodChannel;
 import io.flutter.plugin.common.PluginRegistry;
 
 class FacebookLoginResultDelegate implements FacebookCallback<LoginResult>, PluginRegistry.ActivityResultListener {
     private final CallbackManager callbackManager;
     private MethodChannel.Result pendingResult;
+
+
+    // this int will be used to check if the user has a previous session but with other
+    // user if that is right we will logout and make a login request again
+    private int intent = 0;
+
+
+    public OnReLoginListener onReLoginListener;
 
     FacebookLoginResultDelegate(CallbackManager callbackManager) {
         this.callbackManager = callbackManager;
@@ -36,6 +51,19 @@ class FacebookLoginResultDelegate implements FacebookCallback<LoginResult>, Plug
 
     @Override
     public void onError(FacebookException error) {
+        if (error instanceof FacebookAuthorizationException ) {
+            // if the user has a previous session with other user
+            final boolean prevSessionWithDifferentAccount = AccessToken.getCurrentAccessToken() != null;
+            if (prevSessionWithDifferentAccount && intent == 0) {
+                LoginManager.getInstance().logOut();// close the previous session
+                if (this.onReLoginListener != null) {
+                    this.onReLoginListener.onReLogin();
+                    intent++;
+                    return;
+                }
+            }
+        }
+
         finishWithError("FAILED", error.getMessage());
     }
 
@@ -56,10 +84,9 @@ class FacebookLoginResultDelegate implements FacebookCallback<LoginResult>, Plug
             );
             return false;
         }
-
+        intent = 0;
         pendingResult = result;
         return true;
-
     }
 
 
@@ -69,4 +96,9 @@ class FacebookLoginResultDelegate implements FacebookCallback<LoginResult>, Plug
             pendingResult = null;
         }
     }
+}
+
+
+interface OnReLoginListener {
+    void onReLogin();
 }