+++ facebook_auth_web/ios/Flutter/flutter_export_environment.sh
@@ -1,11 +1,11 @@
 #!/bin/sh
 # This is a generated file; do not edit or check into version control.
-export "FLUTTER_ROOT=/Users/darwin/fvm/versions/2.0.0"
+export "FLUTTER_ROOT=/Users/darwin/development/flutter"
 export "FLUTTER_APPLICATION_PATH=/Users/darwin/Documents/meedu/flutter-facebook-auth/facebook_auth_web"
 export "FLUTTER_TARGET=lib/main.dart"
 export "FLUTTER_BUILD_DIR=build"
 export "SYMROOT=${SOURCE_ROOT}/../build/ios"
-export "FLUTTER_BUILD_NAME=2.0.0"
+export "FLUTTER_BUILD_NAME=2.4.1"
 export "FLUTTER_BUILD_NUMBER=1"
 export "DART_OBFUSCATION=false"
 export "TRACK_WIDGET_CREATION=false"